+++ src/renderer/helpers/colorUtil.ts
@@ -54,3 +54,27 @@ export const hsvToRgb = ({ h, s, v }: HsvColor): RgbColor => {
         b: Math.round(f(1) * 255),
     };
 };
+
+export const parseRgbHex = (hex: string): RgbColor | undefined => {
+    if (hex.startsWith('#')) {
+        hex = hex.slice(1);
+    }
+    if (!/^[0-9a-fA-F]+$/.test(hex)) return undefined;
+    if (hex.length === 3) {
+        const r = Number.parseInt(hex[0], 16) * 17;
+        const g = Number.parseInt(hex[1], 16) * 17;
+        const b = Number.parseInt(hex[2], 16) * 17;
+        return { r, g, b };
+    }
+    if (hex.length === 6) {
+        const r = Number.parseInt(hex.slice(0, 2), 16);
+        const g = Number.parseInt(hex.slice(2, 4), 16);
+        const b = Number.parseInt(hex.slice(4, 6), 16);
+        return { r, g, b };
+    }
+    return undefined;
+};
+export const rgbToHex = ({ r, g, b }: RgbColor): string => {
+    const hex = (r * 256 * 256 + g * 256 + b).toString(16).padStart(6, '0').toUpperCase();
+    return `#${hex}`;
+};
